requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.2
typing_extensions==4.12.2
